requirements.txt
fastapi
pytest
pydantic==1.10.8
starlette
httpx
pytest
mangum
requests
uvicorn
python-dotenv
pypdf
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-multipart
semchunk
tiktoken